Add remaining function definitions
