+++ githubmarkdownui/blocks/leaf.py
@@ -1,36 +1,28 @@
-from typing import List, Optional
+from typing import Optional
 
-from githubmarkdownui.helpers.table import TableAlignment
+
+def thematic_break() -> str:
+    """Returns a <hr> tag, used to create a thematic break. Equivalent to --- in Markdown."""
+    pass
+
+
+def code_block(text: str, language: Optional[str] = None) -> str:
+    """Creates a code block using HTML syntax. If language is given, then the code block will be
+    created using Markdown syntax, but it cannot be used inside a table.
+
+    :param text: The text inside the code block
+    :param language: The language to use for syntax highlighting
+    """
+    pass
 
 
 def heading(text: str, level: int) -> str:
     """Returns a heading with the given text, by wrapping the text with a corresponding <h1> to <h6> tag. 1 is the biggest
     heading while 6 is the smallest heading. The level must be between 1 to 6 inclusive.
 
-    :raise: Exception if the level is not between 1 to 6 inclusive
-    """
-    pass
+    :param text: The text for the heading
+    :param level: The level of the heading between 1 to 6 inclusive
 
-
-def table(content: List[List[str]], alignment: List[Optional[TableAlignment]] = None) -> None:
-    """Creates a table using HTML syntax.
-
-    The content parameter is a list containing lists of equal length, which correspond to the contents of the table.
-    The first sublist is the table headers, and each remaining sublist is one row of the table.
-
-    For example, the content parameter [['column 1', 'column 2'], ['hello', 'world'], ['foo', 'bar']] would create a
-    table like this:
-
-    column 1 | column 2
-    ---------|---------
-    hello    | world
-    foo      | bar
-
-    :param content: A list of lists containing the contents of the table
-    :param alignment: An optional list specifying how each column of the table should be aligned. This list must be the
-    same length as the sublists of the content parameter
-
-    :raises: TableAlignmentError when the alignment parameter is not the same length as the sublists of content
-    :raises: TableDimensionError when the sublists of content do not contain the same number of elements
+    :raises: Exception if the level is not between 1 to 6 inclusive
     """
     pass
