+++ githubmarkdownui/blocks/list.py
@@ -34,3 +34,14 @@ class UnorderedList(HtmlList):
 
     def __str__(self) -> str:
         pass
+
+
+def task_list(items: List[str], items_to_check: List[int]) -> str:
+    """Creates a task list in Markdown where each item can be checked off. This task list cannot be used inside a table.
+
+    :param items: The items in the task list
+    :param items_to_check: The indices (starting from 0) of the items that should be checked off
+
+    :raises: Exception if an index in items_to_check does not exist in the items list
+    """
+    pass
